Implemented some basic libc stuff and added method stubs for other stuff

+++ libc/include/sys/types.h
@@ -1,4 +1,4 @@
 #ifndef _SYS_TYPES_H
 #define _SYS_TYPES_H
-typedef int pid_t;
+typedef unsigned long pid_t;
 #endif
